+++ database/migrations/2020_03_06_151350_create_user_limits_table.php
@@ -18,7 +18,7 @@ class CreateUserLimitsTable extends Migration
             $table->foreignId('user_id')->constrained();
             $table->foreignId('category_id')->constrained();
             $table->integer('limit_per')->nullable();
-            $table->integer('duration');
+            $table->tinyInteger('duration');
             $table->timestamps();
 
             $table->unique(['user_id', 'category_id']);
